+++ apps/server/src/infra/sync/sync.module.ts
@@ -10,6 +10,7 @@ import { LoggerModule } from '@src/core/logger';
 import { ProvisioningModule } from '@src/modules/provisioning';
 import { SyncConsole } from './console/sync.console';
 import { SyncService } from './service/sync.service';
+import { TspOauthDataMapper } from './tsp/tsp-oauth-data.mapper';
 import { TspSyncService } from './tsp/tsp-sync.service';
 import { TspSyncStrategy } from './tsp/tsp-sync.strategy';
 import { SyncUc } from './uc/sync.uc';
@@ -26,7 +27,9 @@ import { SyncUc } from './uc/sync.uc';
 		SyncConsole,
 		SyncUc,
 		SyncService,
-		...((Configuration.get('FEATURE_TSP_SYNC_ENABLED') as boolean) ? [TspSyncStrategy, TspSyncService] : []),
+		...((Configuration.get('FEATURE_TSP_SYNC_ENABLED') as boolean)
+			? [TspSyncStrategy, TspSyncService, TspOauthDataMapper]
+			: []),
 	],
 	exports: [SyncConsole],
 })
